+++ internal/repository/generator.go
@@ -9,11 +9,20 @@ import (
 	"github.com/snapp-incubator/crafting-table/internal/structure"
 )
 
-var createSyntax = ""
-var updateSyntax = ""
-var getSyntax = ""
+func Generate(source, destination, packageName string, getVars *[]structure.Variables, updateVars *[]structure.UpdateVariables, create, test bool) error {
+	createSyntax := ""
+	updateSyntax := ""
+	getSyntax := ""
 
-func Generate(source, destination, packageName string, getVars *[]structure.Variables, updateVars *[]structure.UpdateVariables, create bool) error {
+	createTestSyntax := ""
+	updateTestSyntax := ""
+	getTestSyntax := ""
+
+	var testDestination string
+	if test {
+		testDestination = destination[:len(destination)-3] + "_test.go"
+	}
+
 	s, err := structure.BindStruct(source)
 	if err != nil {
 		err = errors.New(fmt.Sprintf("Error in bindStruct: %s", err.Error()))
@@ -31,6 +40,10 @@ func Generate(source, destination, packageName string, getVars *[]structure.Vari
 			return err
 		}
 		signatures = append(signatures, signature)
+
+		if test {
+			createTestSyntax = createTestFunction(s)
+		}
 	}
 
 	if getVars != nil {
@@ -40,6 +53,10 @@ func Generate(source, destination, packageName string, getVars *[]structure.Vari
 			return err
 		}
 		signatures = append(signatures, signatureList...)
+
+		if test {
+			getTestSyntax = getTestFunction(s, getVars)
+		}
 	}
 
 	if updateVars != nil {
@@ -49,6 +66,11 @@ func Generate(source, destination, packageName string, getVars *[]structure.Vari
 			return err
 		}
 		signatures = append(signatures, signatureList...)
+
+		if test {
+			updateTestSyntax = updateTestFunction(s, updateVars)
+
+		}
 	}
 
 	interfaceSyntax := interfaceCreator(s, signatures)
@@ -66,6 +88,22 @@ func Generate(source, destination, packageName string, getVars *[]structure.Vari
 	if err != nil {
 		err = errors.New(fmt.Sprintf("Error in linter: %s", err.Error()))
 		return err
+	}
+
+	if test {
+		testFileContent := createTestTemplate(s, packageName, createTestSyntax, updateTestSyntax, getTestSyntax)
+
+		err = exportRepository(testFileContent, testDestination)
+		if err != nil {
+			err = errors.New(fmt.Sprintf("Error in writeTestFile: %s", err.Error()))
+			return err
+		}
+
+		err = linter(testDestination)
+		if err != nil {
+			err = errors.New(fmt.Sprintf("Error in linter: %s", err.Error()))
+			return err
+		}
 	}
 
 	return nil